pyproject.toml
[project]
name = "zotero-extractor"
version = "0.1.0"
description = "Automated pipeline: Zotero → SCImago SJR → Google Sheets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyzotero>=1.5.0",
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0", # fast fuzzy matching (drop-in replacement for thefuzz)
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ipykernel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]
